Extract edit post logic into a custom hook
