+++ src/features/post/ui/postModal/components/editPost/EditPostForm.tsx
@@ -1,12 +1,10 @@
 'use client'
 
 import { useEffect, useState } from 'react'
-import { toast } from 'react-toastify'
-import { useEditPostMutation } from '@/features/post/api/postsApi'
 import { Post } from '@/features/post/types/types'
-import { Button, Loader, PublicationDescription } from '@/shared/ui'
 import { User, UserCard } from '@/shared/ui/userCard'
-import { extractMessage, isFetchBaseQueryError } from '@/shared/utils/handleErrors/handleErrors'
+import { Button, PublicationDescription } from '@/shared/ui'
+import { useEditPostHandler } from './useEditPostHandler'
 
 import s from './EditPostForm.module.scss'
 
@@ -21,28 +19,19 @@ type Props = {
 
 export const EditPostForm = ({ user, defaultDescription, postId, onCancelEdit, onSaveEdit, onFormChange }: Props) => {
   const [description, setDescription] = useState(defaultDescription)
-  const [editPost, { isLoading }] = useEditPostMutation()
+  const { handleEditPost, isLoading } = useEditPostHandler()
 
   useEffect(() => {
     const hasChanges = description.trim() !== defaultDescription.trim()
     onFormChange(hasChanges)
   }, [description, defaultDescription, onFormChange])
 
-  const handleSave = async () => {
-    try {
-      const updatedPost = await editPost({ postId, description }).unwrap()
-      onSaveEdit(updatedPost)
-      toast.success('Post updated successfully!')
-    } catch (error) {
-      if (isFetchBaseQueryError(error)) {
-        const message = extractMessage(error.data, 'Failed to update the post.')
-        toast.error(message)
-      } else if (error instanceof Error) {
-        toast.error(error.message)
-      } else {
-        toast.error('An unknown error occurred.')
-      }
-    }
+  const handleSave = () => {
+    handleEditPost({
+      postId,
+      description,
+      onSuccess: onSaveEdit,
+    })
   }
 
   return (
